@@ -2,46 +2,59 @@
 	import { enhance } from '$app/forms';
 	import Icon from '@iconify/svelte';
 	import { toast } from 'svelte-sonner';
-	import { onMount, tick } from 'svelte';
+	import { tick } from 'svelte';
 	import { goto } from '$app/navigation';
 	import * as Dialog from '$lib/components/ui/dialog';
 	import { Button } from '$lib/components/ui/button/index.js';
 	import { formatFriendlyDate, timeSince } from '$lib/utils';
 
-	export let postAuthor;
-	export let postAuthorId;
-	export let postContent;
-	export let comments;
-	export let postDate;
-	export let avatar;
-	export let likes; // actual likes array from parent
-	export let id;
-	export let currentUser;
+	interface Props {
+		postAuthor: any;
+		postAuthorId: any;
+		postContent: any;
+		comments: any;
+		postDate: any;
+		avatar: any;
+		likes: any;
+		id: any;
+		currentUser: any;
+	}
 
-	let isDeleting = false;
+	let {
+		postAuthor,
+		postAuthorId,
+		postContent,
+		comments,
+		postDate,
+		avatar,
+		likes,
+		id,
+		currentUser
+	}: Props = $props();
+
+	let isDeleting = $state(false);
 	let deleteLoading = false;
-	let dialogOpen = false;
+	let dialogOpen = $state(false);
 
-	let isLiked = false;
-	let optimisticLikes: number;
-	let hasOptimisticallyUpdated = false; // Track if the user has clicked the like button
-	let isLocked = false; // Lock to prevent premature syncing
+	let isLiked = $state(false);
+	let optimisticLikes: number = $state(0);
+	let hasOptimisticallyUpdated = $state(false); // Track if the user has clicked the like button
+	let isLocked = $state(false); // Lock to prevent premature syncing
 
 	// Remove the onMount block and use a reactive statement
-	// Set `isLiked` whenever `likes` or `currentUser` changes
-	$: isLiked = likes.includes(currentUser.id);
-	$: optimisticLikes = likes.length; // Set the actual number of likes reactively
 
 	// Set initial likes count and liked state on mount
-	onMount(() => {
+	$effect(() => {
 		isLiked = likes.includes(currentUser.id);
 		optimisticLikes = likes.length; // Set the actual number of likes on mount
 	});
 
 	// Sync with parent data only if not locked and no optimistic update
-	$: if (!hasOptimisticallyUpdated && !isLocked && likes.length !== optimisticLikes) {
-		optimisticLikes = likes.length; // Sync with actual likes only if no optimistic update and not locked
-	}
+	$effect(() => {
+		if (!hasOptimisticallyUpdated && !isLocked && likes.length !== optimisticLikes) {
+			optimisticLikes = likes.length; // Sync with actual likes only if no optimistic update and not locked
+		}
+	});
 
 	// Toggle liked state and update the number of likes optimistically
 	const toggleLiked = () => {
@@ -58,7 +71,7 @@
 
 <div class="text-elipsis relative cursor-pointer overflow-x-hidden transition-all duration-300">
 	<div class="card-body p-3 px-1 transition-all duration-300">
-		<div class="flex items-start gap-2">
+		<div class="flex items-start gap-3">
 			<div>
 				<a href={`/users/${postAuthorId}`}>
 					<div class="h-10 w-10 md:h-12 md:w-12">
@@ -129,7 +142,9 @@
 
 							<button type="submit" class="flex items-center">
 								<Icon
-									icon={isLiked ? 'ph:heart-fill' : 'ph:heart'}
+									icon={isLiked
+										? 'material-symbols:favorite-rounded'
+										: 'material-symbols:favorite-outline-rounded'}
 									class={`h-5 w-5 ${isLiked ? 'text-destructive' : ''}`}
 								/>
 								<span class="sr-only">Like</span>
@@ -140,7 +155,7 @@
 						<div class="font-thin">{optimisticLikes ?? 0}</div>
 					</div>
 
-					<button class="font-thin" on:click={() => goto(`/guestbook/post/${id}`)}>
+					<button class="font-thin" onclick={() => goto(`/guestbook/post/${id}`)}>
 						<div class="flex items-center gap-1">
 							<Icon icon="mdi:comment-outline" class="h-5 w-5" />
 							<div>{comments.length}</div>
