@@ -13,30 +13,34 @@
 	import ScrollIndicator from '$lib/components/ui/ScrollIndicator.svelte';
 	import { animateMainStagger } from '$lib/animations';
 
-	let isSubmitting = false;
-
-	export let data: {
-		user: App.User;
-		posts: App.Post[];
-		post: App.Post;
-	};
-
-	export let form: {
+	interface Props {
 		data: {
-			content?: string;
-			post?: string;
+			user: App.User;
+			posts: App.Post[];
+			post: App.Post;
 		};
-		errors: {
-			content?: string[];
+		form: {
+			data: {
+				content?: string;
+				post?: string;
+			};
+			errors: {
+				content?: string[];
+			};
 		};
-	};
+	}
 
-	let loading = false;
+	let { data, form }: Props = $props();
 
-	$: currentUser.set(data.user);
+	let loading: boolean = $state(false);
+	let isSubmitting: boolean = $state(false);
 
-	let showScrollToTop = false;
-	let sortOption = 'date'; // Default sort option
+	$effect(() => {
+		currentUser.set(data.user);
+	});
+
+	let showScrollToTop = $state(false);
+	let sortOption = $state('date'); // Default sort option
 
 	function sortByDate(posts: App.Post[]) {
 		return posts.sort((a, b) => new Date(b.created).getTime() - new Date(a.created).getTime());
@@ -55,9 +59,11 @@
 		return posts.filter((post) => $currentUser.following.includes(post.author));
 	}
 
-	$: emptyMentioningPostsCount = data.posts.filter((post) => post.mentioning.length === 0).length;
+	let emptyMentioningPostsCount = $derived(
+		data.posts.filter((post) => post.mentioning.length === 0).length
+	);
 
-	$: sortedPosts = (() => {
+	let sortedPosts = $derived.by(() => {
 		let posts = [...data.posts]; // Create a copy of posts to avoid mutation
 		switch (sortOption) {
 			case 'following':
@@ -70,7 +76,7 @@
 			default:
 				return sortByDate(posts);
 		}
-	})();
+	});
 
 	const handleScroll = () => {
 		const shouldShow = window.scrollY > 100;
@@ -174,7 +180,7 @@
 
 <div class="mx-auto w-full max-w-lg transition-all duration-300">
 	<div class="animate-item">
-		<h1 class="flex items-center text-7xl font-bold text-primary">
+		<h1 class="flex items-center text-6xl font-bold text-primary">
 			<span class="title-guest">guest</span>
 			<span class="title-book font-thin text-primary/50">book</span>
 		</h1>
@@ -224,7 +230,7 @@
 								class="text-xs transition-all duration-300"
 							>
 								likes
-								<Icon icon="mdi:heart" class="ml-1 h-4 w-4" />
+								<Icon icon="material-symbols:favorite-rounded" class="ml-1 h-4 w-4" />
 							</Button>
 
 							<Button
@@ -243,7 +249,7 @@
 						{#if sortedPosts.length > 0}
 							{#each sortedPosts as post}
 								{#if post.mentioning.length === 0}
-									<div class="post-wrapper border-b">
+									<div class="border-b">
 										<Post
 											postAuthorId={post.author}
 											comments={post.mentionedBy}
