@@ -1,8 +1,8 @@
-import { pb } from '$lib/pocketbase'
-import { currentUser } from '$lib/stores/user'
+import { pb } from '$lib/pocketbase';
+import { currentUser } from '$lib/stores/user';
 
-pb.authStore.loadFromCookie(document.cookie)
+pb.authStore.loadFromCookie(document.cookie);
 pb.authStore.onChange(() => {
-  currentUser.set(pb.authStore.model)
-  document.cookie = pb.authStore.exportToCookie({ httpOnly: false })
-}, true)
+	currentUser.set(pb.authStore.model);
+	document.cookie = pb.authStore.exportToCookie({ httpOnly: false });
+}, true);
