@@ -11,6 +11,7 @@
 	import { toast } from '$lib/stores/toast';
 
 	export let data: PageData;
+
 	$: currentUser.set(data.user);
 
 	// START VIEW TRANSITIONS API
@@ -40,7 +41,7 @@
 {/if}
 
 <div class="flex min-h-[calc(100svh)] flex-col md:min-h-screen">
-	<Nav />
+	<Nav notifications={data.globalNotifications.length} />
 	<main class="mx-auto my-2 w-full max-w-5xl flex-grow overflow-x-clip px-2 md:my-5">
 		<slot />
 	</main>
